refactor: 카드 추가 버튼과 다시 시도 버튼을 공통 컴포넌트 버튼으로 변경
